Agregar menu de compra

+++ Visual/AnimalVisual.java
@@ -18,7 +18,7 @@ public class AnimalVisual extends Mob {
         this.animal = animal;
         rand = new Random(867);
         this.setRapidez(0.6f);
-        this.posicion = new Punto(rand.nextFloat(habitat.getWidth()), rand.nextFloat(habitat.getHeight()));
+        this.posicion = new Punto(rand.nextInt(habitat.getWidth()), rand.nextInt(habitat.getHeight()));
         //TODO cargar imagen del animal
         imagen = (new ImageIcon("recursos/animales/animal.png")).getImage().getScaledInstance(30,30,Image.SCALE_SMOOTH);
     }
